Celery task create and result getting is now threaded
Calling the celery task used to cause the main twisted thread to block
while it serialized the data, etc. This is now deferred to a thread.

Checking and getting the result of the celery task also used to block
the reactors thread. This has also been deferred to a thread.